fix: abstracts tips
business logic for showing tips like firefox  needed to be applied
in other places such as publishers page.
